Refactoring: extract FileWatcher
