+++ core/src/main/scala/stainless/MainHelpers.scala
@@ -3,11 +3,7 @@
 package stainless
 
 import java.io.{ File, PrintWriter }
-import java.nio.file.{ FileSystems, Path, StandardWatchEventKinds }
-import java.util.concurrent.{ Executors, TimeUnit }
-
-import scala.collection.JavaConversions._
-import scala.collection.mutable.{ Map => MutableMap }
+import java.util.concurrent.Executors
 
 import org.json4s.JsonAST.JObject
 import org.json4s.JsonDSL._
@@ -93,65 +89,16 @@ trait MainHelpers extends inox.MainHelpers {
 
     // When in "watch" mode, no final report is printed as there is no proper end.
     // In fact, we might not even have a full list of VCs to be checked...
-    val watch = ctx.options.findOption(optWatch) getOrElse false
+    val watch = ctx.options.findOptionOrDefault(optWatch)
 
     if (watch) {
-      // TODO wrap this in an utility class.
-
-      // Watch each individual file for modification through their parent directories
-      // (because a WatchService cannot observe files directly..., also we need to keep
-      // track of the modification time because we sometimes receive several event
-      // for the same file...)
-      val watcher = FileSystems.getDefault.newWatchService()
       val files: Set[File] = compiler.sources.toSet map { file: String => new File(file).getAbsoluteFile }
-      val times = MutableMap[File, Long]() ++ (files map { f => f -> f.lastModified })
-      val dirs: Set[Path] = files map { _.getParentFile.toPath }
-      dirs foreach { _.register(watcher, StandardWatchEventKinds.ENTRY_MODIFY) }
-
-      ctx.reporter.info(s"\n\nWaiting for source changes...\n\n")
-
-      var loop = true
-
-      val interruptible = new inox.utils.Interruptible {
-        override def interrupt(): Unit = { loop = false }
+      def action() = {
+        compiler.run()
+        compiler.join()
       }
-      ctx.interruptManager.registerForInterrupts(interruptible)
-
-      while (loop) {
-        // Wait for further changes, filtering out everything that is not of interest
-
-        val key = watcher.poll(500, TimeUnit.MILLISECONDS)
-        if (key != null) {
-          val events = key.pollEvents()
-          val notifications = events map { _.context } collect { case p: Path => p.toAbsolutePath.toFile }
-          val modified = notifications filter files
-
-          // Update the timestamps while looking for things to process.
-          // Note that timestamps are not 100% perfect filters: the files could have the same content.
-          // But it's very lightweight and the rest of the pipeline should be able to handle similar
-          // content anyway.
-          var proceed = false
-          for {
-            f <- modified
-            if f.lastModified > times(f)
-          } {
-            proceed = true
-            times(f) = f.lastModified
-          }
-
-          if (proceed) {
-            ctx.reporter.info(s"Detecting some file modifications...: ${modified mkString ", "}")
-            compiler.run()
-            compiler.join()
-            ctx.reporter.info(s"\n\nWaiting for source changes...\n\n")
-          }
-
-          key.reset()
-        }
-      }
-
-      ctx.interruptManager.unregisterForInterrupts(interruptible)
-      watcher.close()
+      val watcher = new utils.FileWatcher(ctx, files, action)
+      watcher.run()
     } else {
       // Process reports: print summary/export to JSON
       val reports: Seq[AbstractReport] = compiler.getReports
@@ -170,7 +117,9 @@ trait MainHelpers extends inox.MainHelpers {
 
     // Shutdown the pool for a clean exit.
     val unexecuted = MainHelpers.executor.shutdownNow()
-    if (unexecuted.size != 0) ctx.reporter.error("Some tasks were not run (" + unexecuted.size + ")")
+    if (!ctx.interruptManager.isInterrupted && unexecuted.size != 0) {
+      ctx.reporter.error("Some tasks were not run (" + unexecuted.size + ")")
+    }
   } catch {
     case _: inox.FatalError => System.exit(1)
   }
